New Project Requirements for Cat Clicker
Visuals
The application should display two cats. Each cat includes
the cat's name
a picture of the cat
text showing the number of clicks
The specifics of the layout do not matter, so style it however you'd 
like.
Interaction
The number of clicks should increment when each cat picture is clicked.