Project Requirements for Cat Clicker Premium Pro
Visuals
The application should display

a list of cats by name
an area to display the selected cat
an admin button
an admin area with inputs for changing the cat's name, url, and number 
of clicks (hidden by default)
In the cat display area, the following should be displayed

the cat's name
a picture of the cat
text showing the number of clicks
The specifics of the layout do not matter, so style it however you'd 
like.
Interaction
When a cat name is clicked in the list, the cat display area should 
update to show the data for the selected cat.
The number of clicks in the cat area should be unique to each cat, and 
should increment when the cat's picture is clicked.
When the admin button is clicked, the admin area should appear with the 
inputs filled in for the currently-selected cat.
When the cancel button in the admin area is pressed, the admin area 
disappears.
When the save button in the admin area is pressed, the 
currently-selected cat's values update with the values in the admin 
area, and the admin area disappears.